test(cli): mirror credential ownership seams
