+++ packages/pyric-tools/test/credentials/node/from-adc.test.ts
@@ -3,9 +3,7 @@ import { writeFile, rm } from 'node:fs/promises';
 import { tmpdir } from 'node:os';
 import { join } from 'node:path';
 
-import { resolveScope } from '../../src/cli/scope.js';
-import { fromAdc } from '../../src/credentials/node/from-adc.js';
-import type { ProjectScope } from '../../src/credentials/core/types.js';
+import { fromAdc } from '../../../src/credentials/node/from-adc.js';
 
 const realFetch = globalThis.fetch;
 
@@ -50,6 +48,35 @@ describe('fromAdc', () => {
     }
   });
 
+  it('uses the authorized_user fields for the refresh-token exchange', async () => {
+    const path = temporaryAdcPath();
+    await writeFile(
+      path,
+      JSON.stringify({
+        type: 'authorized_user',
+        client_id: 'adc-client',
+        client_secret: 'adc-secret',
+        refresh_token: 'adc-refresh-token',
+      }),
+    );
+    let request: Request | undefined;
+    globalThis.fetch = async (input, init) => {
+      request = new Request(input, init);
+      return new Response(JSON.stringify({ access_token: 'access-token', expires_in: 3600 }));
+    };
+
+    try {
+      const scope = await fromAdc('project', env(), path);
+      await scope?.resolveToken();
+      expect(request?.url).toBe('https://oauth2.googleapis.com/token');
+      expect(await request?.text()).toBe(
+        'client_id=adc-client&client_secret=adc-secret&refresh_token=adc-refresh-token&grant_type=refresh_token',
+      );
+    } finally {
+      await rm(path, { force: true });
+    }
+  });
+
   it('returns null for a missing credential file', async () => {
     expect(await fromAdc('project', env(), temporaryAdcPath())).toBeNull();
   });
@@ -83,59 +110,5 @@ describe('fromAdc', () => {
     } finally {
       await rm(path, { force: true });
     }
-  });
-});
-
-describe('resolveScope', () => {
-  it('prefers a service-account environment credential', async () => {
-    const serviceAccount = Buffer.from(
-      JSON.stringify({
-        client_email: 'sa@example.iam.gserviceaccount.com',
-        private_key: 'private-key',
-        project_id: 'service-account-project',
-      }),
-    ).toString('base64');
-
-    const resolved = await resolveScope({
-      env: env({ FIREBASE_SA_BASE64: serviceAccount }),
-    });
-
-    expect(resolved.source).toBe('FIREBASE_SA_BASE64');
-    expect(resolved.scope.projectId).toBe('service-account-project');
-  });
-
-  it('uses ADC as the final credential source', async () => {
-    const adcScope: ProjectScope = {
-      projectId: 'project',
-      resolveToken: async () => 'adc-token',
-    };
-    const resolved = await resolveScope({
-      env: env(),
-      projectId: 'project',
-      adc: async () => adcScope,
-    });
-
-    expect(resolved.source).toBe('adc');
-    expect(resolved.scope).toBe(adcScope);
-  });
-
-  it('reports the supported Rules Test API credential sources when none resolve', async () => {
-    await expect(
-      resolveScope({
-        env: env(),
-        projectId: 'project',
-        adc: async () => null,
-      }),
-    ).rejects.toThrow('Rules Test API verification requires');
-  });
-
-  it('does not accept a Pyric refresh token as a Rules Test API credential', async () => {
-    await expect(
-      resolveScope({
-        env: env({ PYRIC_REFRESH_TOKEN: 'removed-token-source' }),
-        projectId: 'project',
-        adc: async () => null,
-      }),
-    ).rejects.toThrow('Rules Test API verification requires');
   });
 });
