+++ packages/pyric-tools/test/credentials/node/index.test.ts
@@ -1,0 +1,12 @@
+import { describe, expect, it } from 'bun:test';
+
+import * as nodeCredentials from '../../../src/credentials/node/index.js';
+
+describe('@pyric/cli/credentials/node', () => {
+  it('exports only the sources supported by Rules Test API verification', () => {
+    expect(Object.keys(nodeCredentials).sort()).toEqual([
+      'fromAdc',
+      'fromServiceAccount',
+    ]);
+  });
+});
